+++ mex/extractors/datenkompass/transform.py
@@ -175,7 +175,7 @@ def get_title(item: MergedActivity) -> list[str]:
     return collected_titles
 
 
-def get_vocabulary(
+def get_german_vocabulary(
     entries: list[_VocabularyT],
 ) -> list[str | None]:
     """Get german prefLabel for Vocabularies.
@@ -212,6 +212,21 @@ def get_datenbank(item: MergedBibliographicResource) -> str | None:
     return None
 
 
+def reformat_html_links(string: str) -> str:
+    """Reformat html-formated links in string into plain text urls.
+
+    Args:
+        string: string with possible html-formated links
+
+    Returns:
+        string with reformated plain text urls.
+    """
+    soup_string = BeautifulSoup(string, "html.parser")
+    for a in soup_string.find_all("a", href=True):
+        a.replace_with(a["href"])
+    return str(soup_string)
+
+
 def transform_activities(
     filtered_merged_activities: list[MergedActivity],
     merged_organizational_units_by_id: dict[
@@ -233,11 +248,8 @@ def transform_activities(
     for item in filtered_merged_activities:
         beschreibung = "Es handelt sich um ein Projekt/ Vorhaben. "
         if item.abstract:
-            beschreibung += delim.join(get_german_text(item.abstract))
-            beschreibung_soup = BeautifulSoup(beschreibung, "html.parser")
-            for a in beschreibung_soup.find_all("a", href=True):
-                a.replace_with(a["href"])
-            beschreibung = str(beschreibung_soup)
+            b2 = delim.join(get_german_text(item.abstract))
+            beschreibung += reformat_html_links(b2)
         kontakt = get_email(
             item.responsibleUnit,
             merged_organizational_units_by_id,
@@ -247,7 +259,7 @@ def transform_activities(
             merged_organizational_units_by_id,
         )
         titel = get_title(item)
-        schlagwort = get_vocabulary(item.theme)
+        schlagwort = get_german_vocabulary(item.theme)
         datenbank = []
         if item.website:
             url_de = [w.url for w in item.website if w.language == "de"]
@@ -322,14 +334,11 @@ def transform_bibliographic_resources(
         if len(item.creator) > max_number_authors_cutoff:
             creator_collection += " / et al."
         titel = f"{title_collection} ({creator_collection})"
-        vocab = get_vocabulary(item.bibliographicResourceType)
-        beschreibung = f"{delim.join(s for s in vocab if s is not None)}. "
+        vocab = get_german_vocabulary(item.bibliographicResourceType)
+        beschreibung = f"{delim.join(v for v in vocab if v is not None)}. "
         if item.abstract:
             b2 = delim.join(get_german_text(item.abstract))
-            beschreibung_soup = BeautifulSoup(b2, "html.parser")
-            for a in beschreibung_soup.find_all("a", href=True):
-                a.replace_with(a["href"])
-            beschreibung += str(beschreibung_soup)
+            beschreibung += reformat_html_links(b2)
         datenkompass_bibliographic_recources.append(
             DatenkompassBibliographicResource(
                 beschreibung=beschreibung,
@@ -400,7 +409,7 @@ def transform_resources(
             elif item.accessRestriction == AccessRestriction["OPEN"]:
                 voraussetzungen = "Frei zugänglich"
             frequenz = (
-                get_vocabulary([item.accrualPeriodicity])
+                get_german_vocabulary([item.accrualPeriodicity])
                 if item.accrualPeriodicity
                 else []
             )
@@ -414,17 +423,14 @@ def transform_resources(
             )
             beschreibung = "n/a"
             if item.description:
-                beschreibung = delim.join(get_german_text(item.description))
-                beschreibung_soup = BeautifulSoup(beschreibung, "html.parser")
-                for a in beschreibung_soup.find_all("a", href=True):
-                    a.replace_with(a["href"])
-                beschreibung = str(beschreibung_soup)
+                b2 = delim.join(get_german_text(item.description))
+                beschreibung = reformat_html_links(b2)
             rechtsgrundlagen_benennung = [
                 *[entry.value for entry in item.hasLegalBasis],
-                *get_vocabulary([item.license] if item.license else []),
+                *get_german_vocabulary([item.license] if item.license else []),
             ]
             schlagwort = [
-                *get_vocabulary(item.theme),
+                *get_german_vocabulary(item.theme),
                 *[entry.value for entry in item.keyword],
             ]
             datennutzungszweck = datennutzungszweck_by_primary_source[primary_source]